+++ src/onboarding/handlers/event-handler.ts
@@ -1,7 +1,7 @@
 import * as fs from "fs";
 import * as path from "path";
 import DiscordEvent from "../events/DiscordEvent";
-import { ClientManager } from "../utils/client_manager";
+import { ClientManager } from "../utils/client-manager";
 
 export default async () => {
